Generalize parsing logic to work in more cases
Now, we attempt parsing the RHS of the split as either u32 or IpAddr;
we match on a tuple containing the results from both.

- If the RHS was parsed as a u32, then we assume we're dealing with
  CIDR and needn't continue.

- If the RHS was unable to be parsed as a u32 but could get parsed as
  an IpAddr, just use the result of that parsing as the netmask.

- Finally if neither parse succeeded (Err(_), Err(e)), return the
  AddrParseError from the attempt at parsing as an IpAddr.

This way of doing things is both compact and readable, however it does
require multiple parse attempts to be fully unwound.  This is fine,
though, since they're both really short.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>