NetAddr: Get rid of address field and replace it with memoized
Of course, one could have just passed network addresses and achieved
the same result, but this way I get to just say the entire thing is
done as it should be.

NetAddr represents a _network_ address, so it makes sense for it to
store the _network_ and the _netmask_, instead of something that will
require recalculation of the network address every time you want to
figure it out.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>