Add a function to convert from IpAddr types
I'm assuming that converting requires zero loss, so in this case what
needs to happen is that we use a full f-full prefix to get the minimum
size for the network.  (One address.)

I will shortly implement a grow method that allows for the expansion of
the network by one bit.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>